Move crate::sql::plan -> crate::planner
