+++ src/lib/mod.rs
@@ -3,6 +3,7 @@ pub mod engine;
 pub mod lang;
 pub mod models;
 pub mod parser;
+pub mod planner;
 pub mod pqlir_parser;
 pub mod sql;
 pub mod utils;
