+++ src/lib/sql/mod.rs
@@ -4,7 +4,6 @@ mod eval;
 mod expr;
 mod field;
 mod filter;
-pub mod plan;
 mod proj;
 mod selector;
 mod utils;
@@ -18,7 +17,6 @@ pub use expr::{Expr, Func};
 pub use field::Field;
 pub use field::SourceValue;
 pub use filter::restrict;
-pub use plan::LogicalPlan;
 pub use proj::Proj;
 pub use selector::Selector;
 pub use selector::SelectorNode;
